% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices2D.R
\name{matrices2D}
\alias{matrices2D}
\title{Creates spots matrices}
\usage{
matrices2D(object, ...)
}
\arguments{
\item{object}{S3 object of working TLC}

\item{...}{Additional graphical parameters. (At this time just experimental)}
}
\value{
Returns S3 object with new variable \code{object$spot_matrices} which is a three dimensional matrix (width, height, and pixel intensity values).
}
\description{
Using spots locations and areas this function creates a matrix for each spot.
}
\examples{
\dontrun{
#just makes spot matrices for selected spots of the object
object <- matrices2D(object)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

