% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize2D.R
\name{normalize2D}
\alias{normalize2D}
\title{Normalization of the matrix values}
\usage{
normalize2D(mat)
}
\arguments{
\item{mat}{Matrix of the TLC plate.}
}
\value{
Normalized matrix.
}
\description{
Experimental function. Normalize matrix data.
}
\examples{
\dontrun{
new_mat <- normalize2D(mat)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

