% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picmatrixTIFF.R
\name{picmatrixTIFF}
\alias{picmatrixTIFF}
\title{Converting TIFF to matrix.}
\usage{
picmatrixTIFF(ff, RGB = TRUE, comb = c(0.3, 0.59, 0.11))
}
\arguments{
\item{ff}{TIFF file}

\item{RGB}{Boolean, TRUE - keeps Red, Green and Blue intensities as three matrices. FALSE - using \code{comb} to combine intensities.}

\item{comb}{Vector, combines intensities according to luma. A vector containing three values for R, G and B conversion.}
}
\value{
Returns combined intensities matrix, or separated R, G, B matrices.
}
\description{
Mostly internal function used by \code{createTLC} function. Additional parameters from \code{createTLC} goes there (RGB, comb).
}
\examples{
\dontrun{
#Internal function used by createTLC(...)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

