% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qtlc.R
\name{print.qtlc}
\alias{print.qtlc}
\title{Print Method for \code{qtlc} object}
\usage{
\method{print}{qtlc}(x, ...)
}
\arguments{
\item{x}{S3 object of the working TLC.}

\item{...}{Additional parameters for the \code{print} method}
}
\value{
Prints \code{qtlc} S3 object details
}
\description{
Redefined \code{print} method.
}
\examples{
\dontrun{
print(object)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

