% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.qtlc.R
\name{print.summary.qtlc}
\alias{print.summary.qtlc}
\title{Summary method for \code{qtlc} S3 object}
\usage{
\method{print}{summary.qtlc}(x, ...)
}
\arguments{
\item{x}{S3 object of the working TLC.}

\item{...}{Additional parameters.}
}
\value{
Summary.
}
\description{
Summary method for \code{qtlc} S3 object
}
\examples{
\dontrun{
summary(object)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

