% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotatev.R
\name{rotatev}
\alias{rotatev}
\title{Rotate vertically}
\usage{
rotatev(mat)
}
\arguments{
\item{mat}{The matrix.}
}
\value{
Rotated matrix.
}
\description{
Rotate entire matrix vertically. Mostly internal function.
}
\examples{
\dontrun{
new_mat <- rotatev(mat)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

