% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3D.R
\name{s3D}
\alias{s3D}
\title{Internal function used by showtlc3D}
\usage{
s3D(mat, ogl, grey, ...)
}
\arguments{
\item{mat}{Matrix with x,y,Intensity dimensions}

\item{ogl}{If TLC 3D plot use OpenGL library for fast and interactive 3D plot. (This functionality is based on the \code{rgl} package.) Otherwise the function is based on the plot3D package.}

\item{grey}{Boolean, if TRUE, then tlc is greyscaled. Default value is FALSE.}

\item{...}{Additional graphics parameters.}
}
\value{
None.
}
\description{
Internal function used by showtlc3D
}
\examples{
\dontrun{
#Internal function.
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

