% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2D.R
\name{select2D}
\alias{select2D}
\title{Selects spots areas}
\usage{
select2D(object, w, h, col = "white")
}
\arguments{
\item{object}{S3 object of the working TLC.}

\item{w}{Width of the spot area.}

\item{h}{Height of the spot area.}

\item{col}{Color of the border (default white)}
}
\value{
Return S3 object with new variable \code{object$mat_cell} which is list with "w" and "h" values.
}
\description{
Based on the located spots centers (manualy with mouse and function \code{spot2D}) this function defines spots areas.
}
\examples{
\dontrun{
select2D(object, 80, 50)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

