% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showRf.R
\name{showRf}
\alias{showRf}
\title{Shows Rf on the plot}
\usage{
showRf(object, col = "green", adjust = NULL, cex = 0.6)
}
\arguments{
\item{object}{S3 object of the working TLC}

\item{col}{Color of the lines.}

\item{adjust}{Adjustment for the space of the text. Default value is usualy just OK.}

\item{cex}{A zoom factor for the text.}
}
\value{
None.
}
\description{
Shows prior analysed Rf on the new plot of the 2D matrix.
}
\examples{
\dontrun{
showRf(object)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

