% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showtlc2D.R
\name{showtlc2D}
\alias{showtlc2D}
\title{Show TLC matrix as 2D plot}
\usage{
showtlc2D(object, ...)
}
\arguments{
\item{object}{S3 object of the working TLC}

\item{...}{Additional parameters}
}
\value{
None
}
\description{
Using TLC matrix width, height, and intensity parameters this function plot 2D heatmap of the TLC matrix.
}
\examples{
\dontrun{
showtlc2D(object)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

