% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot2D.R
\name{spot2D}
\alias{spot2D}
\title{Locate spots manually.}
\usage{
spot2D(object, col = "white", ...)
}
\arguments{
\item{object}{S3 object of the working TLC.}

\item{col}{Color of the spot locator (default is white)}

\item{...}{Additional parameters.}
}
\value{
S3 object with 'object$spots' added.
}
\description{
The function should be used after 2D TLC matrix was plotted. After function call, the user should manually locate centers of the spots using mouse. (Left click for locate, right for the end of the process.)
}
\examples{
\dontrun{
print(object)
object <- spot2D(object)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

