% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qtlc.R
\name{summary.qtlc}
\alias{summary.qtlc}
\title{Summary method for \code{qtlc} S3 object}
\usage{
\method{summary}{qtlc}(object, ...)
}
\arguments{
\item{object}{S3 object of the working TLC.}

\item{...}{Additional parameters for the \code{summary} method}
}
\value{
Summary.
}
\description{
Summary method for \code{qtlc} S3 object
}
\examples{
\dontrun{
summary(object)
}

}
\author{
Ivan D. Pavicevic, \email{ivanp84@gmail.com}
}

